requirements.txt
Flask==2.0.1
pymongo==4.0.1
python-dotenv==0.19.1